+++ relecov_core/utils/handling_samples.py
@@ -278,8 +278,8 @@ def get_sample_display_data(sample_id, user):
 
 def get_sample_obj_from_sample_name(sample_name):
     """Return the sample instance from its name"""
-    if Sample.objects.filter(sequencing_sample_id__iexact=sample_name).exists():
-        return Sample.objects.filter(sequencing_sample_id__iexact=sample_name).last()
+    if Sample.objects.filter(sequencing_sample_id__exact=sample_name).exists():
+        return Sample.objects.filter(sequencing_sample_id__exact=sample_name).last()
     return None
