+++ relecov_dashboard/templates/relecov_dashboard/variantsMutationsInLineagesHeatmap.html
@@ -3,7 +3,7 @@
 {% load plotly_dash %}
 {% block content %}
 <br><br><br><br><br><br>
-{% include "relecov_dashboard/variantsSidebar.html"%}
+
 <div class="container-fluid" style="text-align: center;">
     <div class="card d-inline-flex p-2 m-1" style="width: 75%;">
         <div class="card-body border">
