--- relecov_dashboard/templates/relecov_dashboard/variantsMutationsInLineagesTable copy.html
@@ -1,15 +1,0 @@
-{% extends 'relecov_core/base.html' %}
-{% load static %}
-{%load plotly_dash%}
-{% block content %}
-<br><br><br><br><br><br>
-{% include "relecov_dashboard/variantsSidebar.html"%}
-<div class="container-fluid" style="text-align: center;">
-    <div class="card d-inline-flex p-2 m-1" style="width: 75%;">
-        <div class="card-body border">
-            {% plotly_app name="mutation_table" ratio=0.5 %}
-        </div>
-    </div>
-</div>
-
-{% endblock %}
